Compile with PL/Scope: continue if unexpected errors happen
This adds an option to continue if a recompilation statement fails with
an unexpected exception. While some exceptions may predictably happen,
hence have been treated as expected/non-raising so far, in schemas with
residual errors even more exceptions may happen, that may be difficult
to always anticipate. Besides the distinction between a normal, expected
exception and an abnormal, unanticipated one depends on the user's point
of view. In any case, not stopping (i.e. not raising) on unanticipated
exceptions enables to try to recompile as many objects as possible, without
blocking on the first unanticipated (but actually possibly normal) error.
The default choice of the new option is "Continue", but can be changed to
"Raise" if necessary.