Compile with PL/Scope: enable session settings in the new impl.
Formerly, the choice of the recompilation procedure to run was dictated
by whether prior compilation settings had to be reused or not: the new
(home-brewn) procedure would always reuse settings, whereas the legacy
procedure (using dbms_utility.compile_schema) always enforces session
settings. This is relaxed in that it is now possible to use the new
procedure and yet apply session settings if necessary.