Fix: don't set is_used for formal parameters in forward decl.
This fixes the following issue: formal parameters in a forward declaration
of a function or procedure were always flagged as "unused" (is_used = 'NO')
regardless of whether corresponding parameters in the definition of the
function or procedure were actually used or not. The fix consists in not
setting is_used for formal parameters in forward declarations, leaving that
analysis to corresponding definitions.

On the other hand, is_used is now correctly set for declarations of unused
local functions and procedures, including forward declarations.