dd_util, lineage_util: fix non-compliant variable names
This fixes a few variable names which were not compliant with the Trivadis
PL/SQL naming conventions. (*)

Remark: while some PL/SQL coding conventions (e.g. Kyte, Feuerstein, et al.)
use prefixes solely to indicate scoping, e.g. global variables always start
with "g_", local variables with "l_", regardless of type, per Trivadis PL/SQL
naming conventions, while all global variables (not to be confused with
constants, or explicit cursors) always start with "g_", regardless of type,
prefixes of local variables do depend on type: the "l_" prefix is only for
local variables which are not records (prefix: "r_"), not arrays/tables
(prefix: "t_"), and not objects (prefix: "o_"); explicit cursors, which are
not variables, always starts with "c_", regardless of scope, just as constants
always start with "co_", regardless of scope. The precedence of rules (that
the "g_" prefix has precedence over type for global variables, whereas the
"l_" prefix for local variables has not) is not stated explicitly (*), perhaps
for brevity, meanwhile the definition of the PLSCOPE_NAMING view makes that
point clear, only leaving unresolved whether local cursor variables should
use the "l_" or the "c_" prefix--as a result, as of now the regexp for local
variables defaults to "^(l|c)_.*", due to the view not having a special case
for local variables referencing a refcursor.

(*) As of v4.2 of the Trivadis PL/SQL & SQL Coding Guidelines