Fixed wrong default for array/table type regexp in error msg
In the plscope_naming view, and in corresponding reports, the default
values for regular expressions matching compliant names appear in nvl
expressions which have to occur twice: (a) in regexp_like conditions;
and: (b) in corresponding error messages. In the case of the regular
expression matching array/table type declarations, the default value
in the error message was not identical to the actual default value in
the matching regexp_like condition.